refactoring Main
